+++ packages/plugin-layout/src/utils/intl.ts
@@ -1,4 +1,4 @@
-import { useIntl, getLocale } from 'umi';
+import { getIntl, getLocale } from 'umi';
 import zhCN from '../locale/zh-CN';
 import enUS from '../locale/en-US';
 
@@ -15,15 +15,20 @@ export enum LOCALES_ICON {
 export type ILocale = keyof typeof LOCALES;
 
 /** 处理默认 UI 的国际化函数 */
-export function intl({
+export function formatMessage({
   id,
   value = {},
 }: {
   id: string;
   value?: { [key: string]: any };
 }) {
-  const intl = useIntl();
   const localeMessages: { [key: string]: string } =
     getLocale() === 'zh-CN' ? zhCN : enUS;
+
+  if (!getIntl) {
+    return localeMessages[id] || id;
+  }
+  const intl = getIntl();
+
   return intl.formatMessage({ id }, value) || localeMessages[id] || id;
 }