+++ packages/plugin-qiankun/examples/app2/pages/user/index.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import { Button, Table } from 'antd';
 import { connect } from 'dva';
-
+import { connectMaster } from 'umi';
 @connect(({ user }) => ({ user }))
+@connectMaster
 export default class extends React.PureComponent {
   constructor(props) {
     super(props);
@@ -13,7 +14,7 @@ export default class extends React.PureComponent {
   }
 
   render() {
-    const { user } = this.props;
+    const { user, current, pageSize } = this.props;
     const { list } = user;
     const columns = [
       {
@@ -32,6 +33,11 @@ export default class extends React.PureComponent {
 
     return (
       <div>
+        <div>
+          <h1>来自 app1 的数据</h1>
+          <p>current: {current}</p>
+          <p>pageSize: {pageSize}</p>
+        </div>
         <Table rowKey="id" columns={columns} dataSource={list} />
       </div>
     );