+++ packages/plugin-request/tests/index.test.js
@@ -25,6 +25,7 @@ describe('plugin-request', () => {
         },
       },
       addUmiExports() {},
+      addDepInfo() {},
       writeTmpFile,
     };
   };
@@ -49,7 +50,7 @@ describe('plugin-request', () => {
 
     expect(writeTmpFile).toHaveBeenLastCalledWith({
       path: 'plugin-request/request.ts',
-      content: expect.stringContaining('result: T;'),
+      content: expect.stringContaining('result?: T;'),
     });
 
     expect(writeTmpFile).toHaveBeenLastCalledWith({