+++ scripts/release.js
@@ -109,6 +109,9 @@ async function release() {
       ]
         .concat(conventionalGraduate)
         .concat(conventionalPrerelease),
+      {
+        shell: false,
+      },
     );
   }